added rejoin feature if socket gets disconnected in between (yet to be tested)

+++ socketcontrollers/RoomSocketController.ts
@@ -18,7 +18,6 @@ export class RoomSocketController{
 
     public static create( socket : Socket ) {
         socket.on('create' , async(data:any)=>{
-            const userId = socket.id;
             const authData : RoomAuthData = data;
             if(RoomSocketController.verify(authData)){ 
                 if(RoomSocketController.allRoomNames.includes(authData.roomName)){
@@ -27,9 +26,9 @@ export class RoomSocketController{
                     RoomSocketController.allRoomNames.push(authData.roomName);
                     RoomSchema.create({
                         roomName : authData.roomName ,
-                        admin : userId,
+                        admin : authData.userName,
                         pass : await CryptoController.encode(authData.password) ,
-                        users : [ { userName : authData.userName , userId : userId } ]
+                        users : [ authData.userName ]
                     });
                     socket.emit('create-res',{ 
                         jwt : JwtSocketController.create({roomName : authData.roomName }) ,
@@ -51,40 +50,59 @@ export class RoomSocketController{
         socket.on('join',async(data)=>{
             const userId = socket.id;
             const authData : RoomAuthData = data; 
-            if( ! RoomSocketController.verify(authData) ) socket.emit('join-res',{ message : 'Invalid Credentials'});
+            if( ! RoomSocketController.verify(authData) ) return socket.emit('join-res',{ message : 'Invalid Credentials'});
             const room = await RoomSchema.findOne({
                 roomName : authData.roomName
             });
             if(room === null) return socket.emit('join-res',{message : 'Room Not Found'});
+            else if(room.users.includes(authData.userName)) return socket.emit('join-res',{message : 'Username taken by someone in the room'});       
             if(await CryptoController.verify( authData.password , room.pass )){
-                room.users = [ { userName : authData.userName , userId : userId } , ...room.users ];
+                room.users = [ authData.userName , ...room.users ];
                 room.save();
                 socket.join([authData.roomName]);
                 socket.broadcast.to(authData.roomName).emit('userlist',
-                    {users : room.users.map((e:{userName : string , userId : string})=>e.userName)}
+                    {users : room.users}
                 );
                 socket.emit('join-res',{
                     message : 'success', 
                     jwt : JwtSocketController.create({roomName : authData.roomName}) ,
                     files : FileSocketController.allFilesData(authData.roomName) ,
-                    users : room.users.map((e:{userName : string , userId : string})=>e.userName)
+                    users : room.users
                 });
             }else{
                 socket.emit('join-res',{ message : 'Authentication Failed' });
             }
         });
     }
+
+    public static rejoin(socket: Socket){
+        socket.on('rejoin' , async(data:any)=>{
+            const { token,roomName,userName } = data;
+            if(Verifier.isToken(token) && Verifier.isRoom(roomName)&& Verifier.isUser(userName)){
+                // token has already been verified in middleware
+                const room = await RoomSchema.findOne({
+                    roomName : roomName
+                });
+                if(room == null) return socket.emit('rejoin-res',{ message: "room doesn't exists" });
+                if(room.users.includes(userName)){
+                    socket.join([roomName]);
+                    socket.emit('rejoin-res',{ message: 'success'});
+                }else
+                    socket.emit('rejoin-res',{ message: 'can not find user in mentioned room' });
+            }else
+                socket.emit('rejoin-res',{ message: 'Invalid credentials'});
+        });
+    }
     
 
     public static leave( socket : Socket ){
         socket.on('leave',async(data : any)=>{
-            const userId = socket.id;
-            const { roomName } = data;
-            if(roomName){
+            const { roomName,userName } = data;
+            if(roomName && userName){
                 const room = await RoomSchema.findOne({
                     roomName : roomName
                 });
-                if(userId === room.admin){
+                if(userName === room.admin){
                     RoomSocketController.allRoomNames.splice(RoomSocketController.allRoomNames.indexOf(roomName));
                     RoomSchema.deleteMany({
                         roomName : roomName
@@ -95,10 +113,10 @@ export class RoomSocketController{
                     socket.emit('leave-res',{ message : 'success'});
                 }else{
                     room.users = room.users.filter(
-                        (e:{userName : string , userId : string})=>(e.userId !== userId));
+                        (un : string)=>(userName !== un));
                     room.save();
                     socket.broadcast.to(roomName).emit('userlist',
-                        {users : room.users.map((e:{userName : string , userId : string})=>e.userName) }
+                        {users : room.users }
                     );
                     socket.leave(roomName);
                     socket.emit('leave-res',{ message : 'success'});
